+++ lib/lib_cleanex/plugs.ex
@@ -1,4 +1,4 @@
-defmodule Cleanex.Plug.CleanParams do
+defmodule Cleanex.Plugs.ParamsCleaner do
   @moduledoc """
   A plug to clean `conn.params`
 
@@ -6,14 +6,14 @@ defmodule Cleanex.Plug.CleanParams do
 
       pipeline :api do
         ...
-        plug Cleanex.Plug.CleanParams
+        plug Cleanex.Plug.ParamsCleaner
         ...
       end
 
   """
   def init(opts), do: opts
 
-  def call(conn, params) do
+  def call(conn, _) do
     %{conn | params: Cleanex.clean(conn.params)}
   end
 end