+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+0.3.0 (August 3, 2018)
+======================
+
+* Tidy up and improve library
+* Rename plug to "Cleanex.Plugs.ParamsCleaner"
+* Add dialyzer
+* Add plug tests
+* Fix a bug when null bytes and whitespace appear together
+
+
 0.2.0 (July 13, 2018)
 =====================